fix5

+++ WSEcoOZ/src/main/java/so/ObrisiZahtev.java
@@ -23,8 +23,11 @@ public class ObrisiZahtev extends SOAbstract {
     Zahtev zahtev;
     Logger logger = Logger.getLogger(KreirajKorisnika.class);
     StavkaZahteva stavkaZahteva;
+    String tip = "";
+    int id = 0;
 
-    public ObrisiZahtev(OpstiDomenskiObjekat odo) {
+    public ObrisiZahtev(OpstiDomenskiObjekat odo, String tip) {
+        this.tip = tip;
         if (odo instanceof Zahtev) {
             zahtev = (Zahtev) odo;
             stavkaZahteva = null;
@@ -40,18 +43,44 @@ public class ObrisiZahtev extends SOAbstract {
 
     @Override
     public void execute() {
-        if (stavkaZahteva != null) {
-            session.createNativeQuery("delete from stavka_zahteva where redniBroj=:rb and zahtev_zahtevID=:ID")
-                    .setParameter("rb", stavkaZahteva.getRedniBroj())
-                    .setParameter("ID", stavkaZahteva.getZahtev().getZahtevID()).executeUpdate();
+        double sum = 0;
 
-            session.createNativeQuery("delete from zaduzenja where zahtev_zahtevID=:ID")
-                    .setParameter("ID", stavkaZahteva.getZahtev().getZahtevID()).executeUpdate();
+        if (stavkaZahteva != null) {
+            if (tip.equals("obrisi")) {
+                session.createNativeQuery("delete from stavka_zahteva where redniBroj=" + stavkaZahteva.getRedniBroj() + " and zahtev_zahtevID=" + stavkaZahteva.getZahtev().getZahtevID())
+                        .executeUpdate();
+
+            } else {
+                Zahtev found = (Zahtev) session.createNativeQuery("select * from zahtev where zahtevID=" + stavkaZahteva.getZahtev().getZahtevID())
+                        .addEntity(Zahtev.class).list().get(0);
+                for (StavkaZahteva s : found.getListaStavki()) {
+                    sum += s.getKolicina();
+                }
+                session.createNativeQuery("update zahtev set ukupno=" + sum + " where zahtevID=" + stavkaZahteva.getZahtev().getZahtevID())
+                        .executeUpdate();
+            }
         } else if (zahtev != null) {
-            session.delete(zahtev);
-            session.createNativeQuery("delete from zaduzenja where zahtev_zahtevID=:ID")
-                    .setParameter("ID", zahtev.getZahtevID()).executeUpdate();
+            if (tip.equals("obrisi")) {
+
+                session.createNativeQuery("delete from stavka_zahteva zahtev_zahtevID=:ID")
+                        .setParameter("ID", zahtev.getZahtevID()).executeUpdate();
+
+                session.createNativeQuery("delete from zaduzenja where zahtev_zahtevID=:ID")
+                        .setParameter("ID", stavkaZahteva.getZahtev().getZahtevID()).executeUpdate();
+
+                session.delete(zahtev);
+            } else {
+                Zahtev found = (Zahtev) session.createNativeQuery("select * from zahtev where zahtevID=" + zahtev.getZahtevID())
+                        .addEntity(Zahtev.class).list().get(0);
+                for (StavkaZahteva s : found.getListaStavki()) {
+                    sum += s.getKolicina();
+                }
+                session.createNativeQuery("update zahtev set ukupno=" + sum + " where zahtevID=" + zahtev.getZahtevID())
+                        .executeUpdate();
+            }
+
         }
+
         httpStatus = HttpStatus.OK;
 
     }
